Support DeleteWord with <Alt-D>
